+++ tests/unit/TemplateResolver/Template/LogicVariables/Handlers/DefaultHandlers/Common/FirstCharacterInUppercaseHandlerTest.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace ALI\TextTemplate\Tests\TemplateResolver\Template\LogicVariables\Handlers\DefaultHandlers\Common;
+
+use ALI\TextTemplate\TemplateResolver\Template\LogicVariables\HandlerInterface;
+use ALI\TextTemplate\TemplateResolver\Template\LogicVariables\Handlers\DefaultHandlers\Common\FirstCharacterInUppercaseHandler;
+use PHPUnit\Framework\TestCase;
+
+class FirstCharacterInUppercaseHandlerTest extends TestCase
+{
+    public function test()
+    {
+        $handler = new FirstCharacterInUppercaseHandler();
+
+        $dataForCheck = [
+            'Sun' => "Sun",
+            'sun' => "Sun",
+            'SHop' => "SHop",
+            'sHoP' => "SHoP",
+            '-SHoP#' => "-SHoP#",
+            'i̇stanbul' => "İstanbul",
+            'морозиво' => "Морозиво",
+        ];
+        $this->check($dataForCheck, $handler);
+    }
+
+    protected function check(
+        array            $dataForCheck,
+        HandlerInterface $handler
+    ): void
+    {
+        foreach ($dataForCheck as $inputText => $correctResolvedText) {
+            $this->assertEquals($correctResolvedText, $handler->run($inputText, []));
+        }
+    }
+}
